+++ tests/test_utils.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from django.test.testcases import TestCase
 
-from cruds.utils import get_fields
+from cruds import utils
 
 from tests.testapp.models import (
     Author,
@@ -13,5 +13,19 @@ from tests.testapp.models import (
 class TestUtils(TestCase):
 
     def test_get_fields_order(self):
-        res = get_fields(Author, ('birthday', 'name'))
+        res = utils.get_fields(Author, ('birthday', 'name'))
         self.assertEqual(list(res.keys())[0], 'birthday')
+
+    def test_crud_permission_name(self):
+        self.assertEqual(
+            utils.crud_permission_name(Author, utils.ACTION_UPDATE),
+            'testapp.change_author'
+        )
+        self.assertEqual(
+            utils.crud_permission_name(Author, utils.ACTION_CREATE),
+            'testapp.add_author'
+        )
+        self.assertEqual(
+            utils.crud_permission_name(Author, utils.ACTION_LIST),
+            'testapp.list_author'
+        )